pyproject.toml
[project]
name = "satellite-field-segmentation"
version = "0.1.0"
description = "Field boundary segmentation from satellite imagery using ResUNet-A"
requires-python = ">=3.11"

dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "pyyaml>=6.0.1",
    "rasterio>=1.3.9",
    "geopandas>=0.14.3",
    "shapely>=2.0.3",
    "fiona>=1.9.5",
    "pyproj>=3.6.1",
    "pystac-client>=0.7.5",
    "higra>=0.7.0",
    "scikit-image>=0.22.0",
    "albumentations>=1.4.0",
    "scikit-learn>=1.4.1",
    "tensorboard>=2.16.2",
    "matplotlib>=3.8.3",
    "seaborn>=0.13.2",
    "jupyter>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
